add loader card, moved ui to top level, add file breakdown loaders
