+++ src/components/CohortBuilder/LoaderCard.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { withTheme } from 'emotion-theming';
+import { CohortCard } from './ui';
+import LoadingSpinner from 'uikit/LoadingSpinner';
+
+const LoaderCard = ({ title, theme }) => (
+  <CohortCard title={title}>
+    <LoadingSpinner color={theme.greyScale11} size={'50px'} />
+  </CohortCard>
+);
+
+export default withTheme(LoaderCard);
